+++ src/index.js
@@ -1,46 +1,27 @@
-import CarFactory from "./domain/carfacotory.js";
+import RacingCarGame from "./controller/racing-game-controller.js"
 import { Element, ElementControl } from "./view/element.js";
 
-export default class RacingCarGame {
+export default class App {
   constructor() {
     ElementControl.hideCarGameCountContainer();
     ElementControl.hideCarGameResultContainer();
 
-    this.clickCarNameSubmitBtn();
+    this.setEventListener();
   }
 
-  clickCarNameSubmitBtn() {
+  setEventListener() {
+    const _racingCarGame = new RacingCarGame();
     const _carNamesSubmitElement = Element.carNamesSubmit;
-
-    _carNamesSubmitElement.addEventListener("click", () => {
-      this.handleCarNameInput()
-    });
-  }
-
-  clickRacingCountSubmitBtn() {
     const _racingCountSubmitElement = Element.racingCountSubmit;
 
-    _racingCountSubmitElement.addEventListener("click", () => {
-      this.handleRacingCountInput()
+    _carNamesSubmitElement.addEventListener("click", () => {
+      _racingCarGame.handleCarNameInput()
     });
 
+    _racingCountSubmitElement.addEventListener("click", () => {
+      _racingCarGame.handleRacingCountInput()
+    });
   }
-
-  handleCarNameInput() {
-    const _carFactory = new CarFactory();
-    this.carArray = _carFactory.getCarArray();
-
-    if (this.carArray.length !== 0) {
-      ElementControl.showCarGameCountContainer();
-
-      this.clickRacingCountSubmitBtn();
-    }
-  }
-
-  handleRacingCountInput() {
-    
-  }
-
 }
 
-new RacingCarGame();
+new App();
