+++ src/view/element.js
@@ -2,7 +2,8 @@ export const Element = {
   carGameCountContainer: document.querySelector(".car-game-count-container"),
   carGameResultContainer: document.querySelector(".car-game-result-container"),
   carNamesSubmit: document.querySelector("#car-names-submit"),
-  racingCountSubmit: document.querySelector("#racing-count-submit")
+  racingCountInput: document.querySelector("#racing-count-input"),
+  racingCountSubmit: document.querySelector("#racing-count-submit"),
 };
 
 export const ElementControl = {
